backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
jupyter-client==8.6.0
ipykernel==6.27.1
nbformat==5.9.2
google-generativeai==0.3.2
openai==1.54.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
